src/main/java/hello/servlet/domain/member/MemberRepository.java
package hello.servlet.domain.member;



 ///동시성 문제가 고려되어 있지 않음, 실무에서는 ConcurrentHashMap, Atomiclong 사용 고려

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

public class MemberRepository {
    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    private static final MemberRepository instance = new MemberRepository();

    public static MemberRepository getInstance(){
        return instance;
    }

    private MemberRepository(){
        //생성자는 private을 걸어 아무나 생성하지 못하게한다
    }

    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    public Member findById(Long id){ //아이디로찾기
        return store.get(id);
    }

    public List<Member> findAll(){
        return new ArrayList<>(store.values());
    }
    public void clearStore(){
        store.clear();
    }





}
